+++ tests/unit/Runner/StepRunnerTest.php
@@ -567,6 +567,7 @@ class StepRunnerTest extends RunnerTestCase
     /**
      * @see StepRunner::obtainServicesNetwork()
      * @see StepRunner::shutdownServices()
+     * @covers \Ktomk\Pipelines\Runner\StepContainer::execRunServiceContainer
      */
     public function testServicesObtainNetworkAndShutdown()
     {